requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
playwright==1.48.0
websockets==12.0
sendgrid==6.11.0
python-dotenv==1.0.0
africastalking
redis==5.0.8
sqlalchemy==2.0.36
asyncpg==0.30.0

